Move Parser and Result into new parsing package
