+++ src/main/java/command/Command.java
@@ -1,8 +1,8 @@
 package command;
 
 import io.Storage;
+import parsing.Parser;
 import task.TaskList;
-import util.Parser;
 
 /**
  * Interface for how a command will interact with data and output.
