+++ src/main/java/command/Delete.java
@@ -2,8 +2,8 @@ package command;
 
 import io.Storage;
 import io.Ui;
+import parsing.Parser;
 import task.TaskList;
-import util.Parser;
 
 /**
  * Manages deletion of tasks from tasks.
