+++ src/main/java/command/Exit.java
@@ -2,8 +2,8 @@ package command;
 
 import io.Storage;
 import io.Ui;
+import parsing.Parser;
 import task.TaskList;
-import util.Parser;
 
 public class Exit implements Command {
