+++ src/main/java/command/Mark.java
@@ -2,8 +2,8 @@ package command;
 
 import io.Storage;
 import io.Ui;
+import parsing.Parser;
 import task.TaskList;
-import util.Parser;
 
 public class Mark implements Command {
     private final int taskNum;
